+++ src/test/resources/jp/co/tis/gsp/tools/dba/mojo/GenerateEntity_test/entityTemplate/h2/expected/output/jp/co/tis/gsptest/entity/entity/IndexTest3.java
@@ -36,7 +36,7 @@ public class IndexTest3 implements Serializable {
 // ENTITY_TEMPLATE_TEST!
     @Id
     @GeneratedValue(generator = "generator", strategy = GenerationType.AUTO)
-    @Column(name = "INDEX_TEST3_ID", precision = 19, nullable = false, unique = true)
+    @Column(name = "INDEX_TEST3_ID", precision = 64, nullable = false, unique = true)
     public Long getIndexTest3Id() {
         return indexTest3Id;
     }
@@ -55,7 +55,7 @@ public class IndexTest3 implements Serializable {
      * @return SUB_ID_1
      */
 // ENTITY_TEMPLATE_TEST!
-    @Column(name = "SUB_ID_1", precision = 19, nullable = false, unique = false)
+    @Column(name = "SUB_ID_1", precision = 64, nullable = false, unique = false)
     public Long getSubId1() {
         return subId1;
     }
@@ -74,7 +74,7 @@ public class IndexTest3 implements Serializable {
      * @return SUB_ID_2
      */
 // ENTITY_TEMPLATE_TEST!
-    @Column(name = "SUB_ID_2", precision = 19, nullable = false, unique = false)
+    @Column(name = "SUB_ID_2", precision = 64, nullable = false, unique = false)
     public Long getSubId2() {
         return subId2;
     }
